fix: remove ad error, and mock fragment for storybook (#1303)

+++ packages/ad/src/utils/ad-init.js
@@ -153,7 +153,11 @@ const adInit = args => {
       },
 
       destroySlots() {
-        window.googletag.destroySlots();
+        if (window.googletag.destroySlots) {
+          window.googletag.destroySlots();
+          return true;
+        }
+        return false;
       }
     },
